Form interface changed

+++ numberField.go
@@ -64,7 +64,7 @@ func (f *numberField) Assign(val interface{}) error {
 		}
 		f.value = converted
 	case float64:
-		f.value = int(value)
+		f.value = int64(value)
 	default:
 		return typeMismatchError
 	}
